requirements.txt
pandas>=1.5.0
pyyaml>=6.0
apache-airflow>=2.7.0
pytest>=7.0.0
